bgutil daemon: replace test by a more gine grained one
